Add EmailAccountViewModelService to the admin

+++ Grand.Web/Areas/Admin/Controllers/EmailAccountController.cs
@@ -9,6 +9,7 @@ using Grand.Services.Messages;
 using Grand.Services.Security;
 using Grand.Web.Areas.Admin.Extensions;
 using Grand.Web.Areas.Admin.Models.Messages;
+using Grand.Web.Areas.Admin.Services;
 using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Linq;
@@ -16,39 +17,36 @@ using System.Linq;
 namespace Grand.Web.Areas.Admin.Controllers
 {
     public partial class EmailAccountController : BaseAdminController
-	{
+    {
+        private readonly IEmailAccountViewModelService _emailAccountViewModelService;
         private readonly IEmailAccountService _emailAccountService;
         private readonly ILocalizationService _localizationService;
         private readonly ISettingService _settingService;
-        private readonly IEmailSender _emailSender;
-        private readonly IStoreContext _storeContext;
         private readonly EmailAccountSettings _emailAccountSettings;
         private readonly IPermissionService _permissionService;
 
-		public EmailAccountController(IEmailAccountService emailAccountService,
+        public EmailAccountController(IEmailAccountViewModelService emailAccountViewModelService, IEmailAccountService emailAccountService,
             ILocalizationService localizationService, ISettingService settingService,
-            IEmailSender emailSender, IStoreContext storeContext,
             EmailAccountSettings emailAccountSettings, IPermissionService permissionService)
-		{
+        {
+            this._emailAccountViewModelService = emailAccountViewModelService;
             this._emailAccountService = emailAccountService;
             this._localizationService = localizationService;
             this._emailAccountSettings = emailAccountSettings;
-            this._emailSender = emailSender;
             this._settingService = settingService;
-            this._storeContext = storeContext;
             this._permissionService = permissionService;
-		}
+        }
 
-		public IActionResult List()
+        public IActionResult List()
         {
             if (!_permissionService.Authorize(StandardPermissionProvider.ManageEmailAccounts))
                 return AccessDeniedView();
 
-			return View();
-		}
+            return View();
+        }
 
-		[HttpPost]
-		public IActionResult List(DataSourceRequest command)
+        [HttpPost]
+        public IActionResult List(DataSourceRequest command)
         {
             if (!_permissionService.Authorize(StandardPermissionProvider.ManageEmailAccounts))
                 return AccessDeniedView();
@@ -82,50 +80,43 @@ namespace Grand.Web.Areas.Admin.Controllers
             return RedirectToAction("List");
         }
 
-		public IActionResult Create()
+        public IActionResult Create()
         {
             if (!_permissionService.Authorize(StandardPermissionProvider.ManageEmailAccounts))
                 return AccessDeniedView();
 
-            var model = new EmailAccountModel();
-            //default values
-            model.Port = 25;
-			return View(model);
-		}
+            var model = _emailAccountViewModelService.PrepareEmailAccountModel();
+            return View(model);
+        }
 
         [HttpPost, ParameterBasedOnFormName("save-continue", "continueEditing")]
-		public IActionResult Create(EmailAccountModel model, bool continueEditing)
+        public IActionResult Create(EmailAccountModel model, bool continueEditing)
         {
             if (!_permissionService.Authorize(StandardPermissionProvider.ManageEmailAccounts))
                 return AccessDeniedView();
 
             if (ModelState.IsValid)
             {
-                var emailAccount = model.ToEntity();
-                //set password manually
-                emailAccount.Password = model.Password;
-                _emailAccountService.InsertEmailAccount(emailAccount);
-
+                var emailAccount = _emailAccountViewModelService.InsertEmailAccountModel(model);
                 SuccessNotification(_localizationService.GetResource("Admin.Configuration.EmailAccounts.Added"));
                 return continueEditing ? RedirectToAction("Edit", new { id = emailAccount.Id }) : RedirectToAction("List");
             }
-
             //If we got this far, something failed, redisplay form
             return View(model);
-		}
+        }
 
-		public IActionResult Edit(string id)
+        public IActionResult Edit(string id)
         {
             if (!_permissionService.Authorize(StandardPermissionProvider.ManageEmailAccounts))
                 return AccessDeniedView();
 
-			var emailAccount = _emailAccountService.GetEmailAccountById(id);
+            var emailAccount = _emailAccountService.GetEmailAccountById(id);
             if (emailAccount == null)
                 //No email account found with the specified id
                 return RedirectToAction("List");
 
-			return View(emailAccount.ToModel());
-		}
+            return View(emailAccount.ToModel());
+        }
 
         [HttpPost, ParameterBasedOnFormName("save-continue", "continueEditing")]
         [FormValueRequired("save", "save-continue")]
@@ -141,16 +132,14 @@ namespace Grand.Web.Areas.Admin.Controllers
 
             if (ModelState.IsValid)
             {
-                emailAccount = model.ToEntity(emailAccount);
-                _emailAccountService.UpdateEmailAccount(emailAccount);
-
+                emailAccount = _emailAccountViewModelService.UpdateEmailAccountModel(emailAccount, model);
                 SuccessNotification(_localizationService.GetResource("Admin.Configuration.EmailAccounts.Updated"));
                 return continueEditing ? RedirectToAction("Edit", new { id = emailAccount.Id }) : RedirectToAction("List");
             }
 
             //If we got this far, something failed, redisplay form
             return View(model);
-		}
+        }
 
         [HttpPost, ActionName("Edit")]
         [FormValueRequired("changepassword")]
@@ -163,14 +152,18 @@ namespace Grand.Web.Areas.Admin.Controllers
             if (emailAccount == null)
                 //No email account found with the specified id
                 return RedirectToAction("List");
+            if (ModelState.IsValid)
+            {
+                //do not validate model
+                _emailAccountViewModelService.ChangePasswordEmailAccountModel(emailAccount, model);
+                SuccessNotification(_localizationService.GetResource("Admin.Configuration.EmailAccounts.Fields.Password.PasswordChanged"));
+            }
+            else
+                ErrorNotification(ModelState);
 
-            //do not validate model
-            emailAccount.Password = model.Password;
-            _emailAccountService.UpdateEmailAccount(emailAccount);
-            SuccessNotification(_localizationService.GetResource("Admin.Configuration.EmailAccounts.Fields.Password.PasswordChanged"));
             return RedirectToAction("Edit", new { id = emailAccount.Id });
         }
-        
+
         [HttpPost, ActionName("Edit")]
         [FormValueRequired("sendtestemail")]
         public IActionResult SendTestEmail(EmailAccountModel model)
@@ -182,16 +175,17 @@ namespace Grand.Web.Areas.Admin.Controllers
             if (emailAccount == null)
                 //No email account found with the specified id
                 return RedirectToAction("List");
-
             try
             {
                 if (String.IsNullOrWhiteSpace(model.SendTestEmailTo))
                     throw new GrandException("Enter test email address");
-
-                string subject = _storeContext.CurrentStore.Name + ". Testing email functionality.";
-                string body = "Email works fine.";
-                _emailSender.SendEmail(emailAccount, subject, body, emailAccount.Email, emailAccount.DisplayName, model.SendTestEmailTo, null);
-                SuccessNotification(_localizationService.GetResource("Admin.Configuration.EmailAccounts.SendTestEmail.Success"), false);
+                if (ModelState.IsValid)
+                {
+                    _emailAccountViewModelService.SenTestEmail(emailAccount, model);
+                    SuccessNotification(_localizationService.GetResource("Admin.Configuration.EmailAccounts.SendTestEmail.Success"), false);
+                }
+                else
+                    ErrorNotification(ModelState);
             }
             catch (Exception exc)
             {
@@ -202,30 +196,33 @@ namespace Grand.Web.Areas.Admin.Controllers
             return View(model);
         }
 
-	    [HttpPost]
-	    public IActionResult Delete(string id)
-	    {
-	        if (!_permissionService.Authorize(StandardPermissionProvider.ManageEmailAccounts))
-	            return AccessDeniedView();
+        [HttpPost]
+        public IActionResult Delete(string id)
+        {
+            if (!_permissionService.Authorize(StandardPermissionProvider.ManageEmailAccounts))
+                return AccessDeniedView();
 
-	        var emailAccount = _emailAccountService.GetEmailAccountById(id);
-	        if (emailAccount == null)
-	            //No email account found with the specified id
-	            return RedirectToAction("List");
-
-	        try
-	        {
-	            _emailAccountService.DeleteEmailAccount(emailAccount);
-
-	            SuccessNotification(_localizationService.GetResource("Admin.Configuration.EmailAccounts.Deleted"));
+            var emailAccount = _emailAccountService.GetEmailAccountById(id);
+            if (emailAccount == null)
+                //No email account found with the specified id
+                return RedirectToAction("List");
+            try
+            {
+                if (ModelState.IsValid)
+                {
+                    _emailAccountService.DeleteEmailAccount(emailAccount);
+                    SuccessNotification(_localizationService.GetResource("Admin.Configuration.EmailAccounts.Deleted"));
+                }
+                else
+                    ErrorNotification(ModelState);
 
                 return RedirectToAction("List");
-	        }
-	        catch (Exception exc)
-	        {
-	            ErrorNotification(exc);
-	            return RedirectToAction("Edit", new {id = emailAccount.Id});
-	        }
-	    }
-	}
+            }
+            catch (Exception exc)
+            {
+                ErrorNotification(exc);
+                return RedirectToAction("Edit", new { id = emailAccount.Id });
+            }
+        }
+    }
 }
